+++ src/logic/data-management/SearchTree.ts
@@ -49,7 +49,6 @@ function constructSearchtree(t, il, rerootCallback = x => {}) {
 
         // Add the nested unordered lists to the page based on the data array
         tree = d3Select.select("#searchtree");
-        console.log(tree);
         tree = tree.append("ul").append("li").attr("class", "root not").append("ul");
         // $("li.root").prepend($("#treeSearchDiv"));
         items = tree.selectAll("li").data([data])
