--- src/test/resources/temp.json
@@ -1,45 +1,0 @@
-{
-    "ecnumbers": [
-        "3.4.13.19",
-        "3.4.13.-",
-        "3.4.-.-",
-        "3.-.-.-",
-        "3.4.17.2",
-        "3.4.17.-",
-        "2.7.7.6",
-        "2.7.7.-",
-        "2.7.-.-",
-        "2.-.-.-",
-        "1.4.1.4",
-        "1.4.1.-",
-        "1.4.-.-",
-        "1.-.-.-",
-        "4.1.1.49",
-        "4.1.1.-",
-        "4.1.-.-",
-        "4.-.-.-",
-        "3.4.11.9",
-        "3.4.11.-",
-        "3.4.21.71",
-        "3.4.21.-",
-        "3.5.1.23",
-        "3.5.1.-",
-        "3.5.-.-",
-        "3.2.2.23",
-        "3.2.2.-",
-        "3.2.-.-",
-        "2.5.1.90",
-        "2.5.1.-",
-        "2.5.-.-",
-        "2.7.8.7",
-        "2.7.8.-",
-        "3.6.3.-",
-        "3.6.-.-",
-        "4.2.99.18",
-        "4.2.99.-",
-        "4.2.-.-",
-        "7.1.1.-",
-        "7.1.-.-",
-        "7.-.-.-"
-    ]
-}
